+++ ecaudit/src/test/java/com/ericsson/bss/cassandra/ecaudit/filter/TestDefaultAuditFilter.java
@@ -15,15 +15,9 @@
  */
 package com.ericsson.bss.cassandra.ecaudit.filter;
 
-import java.net.URL;
-import java.util.Properties;
-import java.util.stream.Collectors;
-
 import org.junit.Test;
 
 import com.ericsson.bss.cassandra.ecaudit.entry.AuditEntry;
-import com.ericsson.bss.cassandra.ecaudit.filter.yaml.AuditConfig;
-import com.ericsson.bss.cassandra.ecaudit.filter.yaml.AuditYamlConfigurationLoader;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -38,34 +32,16 @@ public class TestDefaultAuditFilter
     @Test
     public void testWhitelistNotFilteredWithDefaultFilter()
     {
-        Properties properties = getProperties();
-        AuditYamlConfigurationLoader loader = AuditYamlConfigurationLoader.withProperties(properties);
-
-        AuditConfig loadedConfig = loader.loadConfig();
-        assertThat(loadedConfig.getWhitelist()).containsOnly("User1", "User2");
-
         DefaultAuditFilter filter = new DefaultAuditFilter();
 
-        assertThat(loadedConfig.getWhitelist().stream()
-                .map(TestDefaultAuditFilter::toLogEntry)
-                .map(filter::isFiltered)
-                .collect(Collectors.toList()))
-                        .containsExactly(false, false);
-    }
-
-    private static Properties getProperties()
-    {
-        URL url = TestDefaultAuditFilter.class.getResource("/mock_configuration.yaml");
-        Properties properties = new Properties();
-        properties.put(AuditYamlConfigurationLoader.PROPERTY_CONFIG_FILE, url.getPath());
-
-        return properties;
+        assertThat(filter.isFiltered(toLogEntry("user1"))).isFalse();
+        assertThat(filter.isFiltered(toLogEntry("user2"))).isFalse();
     }
 
     private static AuditEntry toLogEntry(String user)
     {
         return AuditEntry.newBuilder()
-                .user(user)
-                .build();
+                         .user(user)
+                         .build();
     }
 }